Fixed crash on game start in Windows.

+++ bot/weapon/UtilityToolBuilder.h
@@ -4,12 +4,12 @@
 
 class UtilityToolBuilder: public WeaponBuilder {
 public:
-	UtilityToolBuilder(int maxRange = 0) :
+	UtilityToolBuilder(float maxRange = 0.0f) :
 			maxRange(maxRange) {
 	}
 
 	std::shared_ptr<Weapon> build(edict_t* weap) const;
 
 private:
-	int maxRange;
+	float maxRange;
 };
